Reorganization of the cogs folder
Réorganisation des fichiers dans ./cogs pour une meilleur lisibilité

./cogs
├───commands
│   ├───moderation
│   │   └───ban
├───database
└───listeners
   └───messages
